refactor finalize_block PoS updates
